Remove deepEqual
