+++ src/assertions/_assertions.js
@@ -3,7 +3,6 @@ import { assertNotEqual } from "./assertNotEqual.js";
 import { assertTrue } from "./assertTrue.js";
 import { assertFalse } from "./assertFalse.js";
 import { assertThrows } from "./assertThrows.js";
-import { assertDeepEqual } from "./assertDeepEqual.js";
 
 
 export default {
@@ -11,6 +10,5 @@ export default {
     assertNotEqual,
     assertTrue,
     assertFalse,
-    assertThrows,
-    assertDeepEqual
+    assertThrows
 }
