+++ tests/example.test.mjs
@@ -1,4 +1,4 @@
-import { describe, it, assertEqual, assertNotEqual, assertTrue, assertFalse, assertThrows, assertDeepEqual } from '../src/testRunner.js';
+import { describe, it, assertEqual, assertNotEqual, assertTrue, assertFalse, assertThrows } from '../src/testRunner.js';
 
 import { helloworld } from '../example/example.mjs';
 
@@ -11,7 +11,7 @@ describe('I am an Example Test Suite', () => {
   helloCommon();
 
   it('should pass equality assertion', () => {
-    assertEqual(1, 2, 'Expected 1 to equal 1');
+    assertEqual(1, 1, 'Expected 1 to equal 1');
   });
 
   it('should pass inequality assertion', () => {
@@ -34,8 +34,8 @@ describe('I am an Example Test Suite', () => {
 
   it('should deeply compare objects', () => {
     const obj1 = { a: 1, b: { c: 2 } };
-    const obj2 = { a: 1, b: { d: 2 } };
-    assertDeepEqual(obj1, obj2, 'Expected objects to be deeply equal');
+    const obj2 = { a: 1, b: { c: 2 } };
+    assertEqual(obj1, obj2, 'Expected objects to be deeply equal');
   });
 
 });
